closer...

+++ tensorRT/build_trt_engine.py
@@ -93,9 +93,9 @@ def _extract_model_inputs_from_batch(batch_data):
         processed['voxel_features'],
         processed['voxel_coords'],
         processed['voxel_num_points'],
-        ego['record_len'],
+        ego['record_len'].to(torch.int32),
         ego['prior_encoding'],
-        ego['spatial_correction_matrix'],
+        ego['spatial_correction_matrix'].to(torch.float32),
     )
